helper-cli: Move extension utility functions to a conventional `Extensions.kt`
